[PLAT-18344][xCluster] Indicate that bootstrapping is always required for automatic mode
Summary:
This diff updates the need_bootstrap API to always indicate that bootstrapping is required for DR configs with automatic ddl replication.

It will need an UI update so the UI passes `isDrConfig` to this API.

Test Plan: - Called the need_bootstrap API with `isDrConfig=true` and made sure the proper response is returned.

Reviewers: amindrov, spothuraju, jmak, nbhatia

Reviewed By: amindrov

Differential Revision: https://phorge.dev.yugabyte.com/D46257