[BACKPORT 2025.1][#27853] docdb: Fix aborted transaction at promotion time leading to stuck Perform
Summary:
In the case where a transaction is aborted by the tserver while we are promoting, but
before the first heartbeat to the new status tablet, we have two cases:
1. heartbeat to the old status tablet is informed first about the abort
2. heartbeat to the new status tablet is informed first about the abort

In case (2), we correctly handle this by notifying waiters on error during CREATED/PROMOTED
heartbeat. But in case (1), the heartbeat is RUNNING (to old status tablet), so we only set error
and do not notify.

If the PROMOTED heartbeat to the new status tablet has already been sent and we are only waiting for
a response, this is OK -- it will get caught once that heartbeat response is returned. But if the
PROMOTED heartbeat has not been sent yet, it will not be sent at all, resulting in a transaction
where no heartbeats are being sent but callbacks are not called, leading to a stuck Perform.

This change fixes the bug by notifying waiters in the case of TransactionStatus::RUNNING heartbeat
under TransactionState::kPromoting.

Waits for GeoTransactionsPromotionConflictAbortTest.TestConflictAbortBeforeNewHeartbeat
which was not supposed to test this case, but ended up doing so in some cases in TSAN, were also
adjusted and a new test
GeoTransactionsPromotionConflictAbortTest.TestConflictAbortViaOldHeartbeatBeforeNewHeartbeat
was added to test this case explicitly.
Jira: DB-17456

Original commit: 78194bd05be8b3a12fc7776889e88f71c8e3e992 / D46138

Test Plan:
`./yb_build.sh --cxx-test pgwrapper_geo_transactions_promotion-test --gtest_filter GeoTransactionsPromotionConflictAbortTest.TestConflictAbortBeforeNewHeartbeat`
`./yb_build.sh --cxx-test pgwrapper_geo_transactions_promotion-test --gtest_filter GeoTransactionsPromotionConflictAbortTest.TestConflictAbortViaOldHeartbeatBeforeNewHeartbeat`

Reviewers: bkolagani

Reviewed By: bkolagani

Subscribers: ybase, yql, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D46296