[#16209] [#27314] YSQL: Add support for specifying colocation relopt on parent partition tables
Summary:
## Motivation

In a colocated db, certain tables can be non-colocated - this allows certain large tables in a db to be sharded as usual while most tables are colocated. For (parent) partitioned tables, this is not supported, however - `create table parentpart(id int) partition by range (id) with (colocation=false);`. This is typically not a problem because (child) partition tables that contain the actual data can be opted out of colocation.

However, in the case where the parent partition is intended to have a hash partitioned primary key, it cannot be created as a colocated table because hash partitioning + colocation is not allowed. Instead, the user has to use a workaround to create the parent partition without a primary key and the child partitions as hash partitioned, which can lead to potential complications with query planning. Instead, fixing the original issue of not being able to opt parent partitions out of colocation is better.

Separately, parent partition tables have corresponding docdb tables due to historical reasons. The number of tablets is calculated like for all other tables, though these docdb tables do not hold any data. This creates unnecessary tablets that matter when there are lot of partition parents.

## Change

1. Support reloption colocation for (parent) partitioned tables, so that `create table parentpart(id int) partition by range (id) with (colocation=false);` works. This reloption setting on the parent partitioned table does not affect the colocation reloption for child partitions, they have to be independently specified if needed.

2. Split parent partition tables into a single tablet always. SPLIT INTO is not supported for parent partition tables, so we don't have to consider overrides to this value from that case.
Jira: DB-5638, DB-16817

Test Plan: Added new test cases.

Reviewers: kfranz, yguan

Reviewed By: kfranz

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D46328