[#28069] YSQL: Avoid multiple calls of PgTablesQueryListener::Ready method
Summary:
The `Waiter` helper class in the `pg_table_cache.cc` file was introduced in the  https://phorge.dev.yugabyte.com/D45542 diff. The intention of this class is to forward the `Result<TableInfo>` on completion of table open RPC to the `TablesResultBuilder` and notify listener with the fully completed `PgTablesQueryResult` object once the builder will build it.

The client code might request multiple tables to be opened at once. In this case  the `PgTablesQueryResult` result object will contains all these table. In this case builder collects all the individual results for each table open operation (the `Result<TableInfo>` objects) tills its count will much the expected number of queried tables.
But in case of failure of opening at least one table the result is treated as completed with error immediately without waiting for other tables. This causes the situation when listener might be notified multiple times.

- suppose table has requested 3 tables to be opened `t_1`, `t_2`, `t_3`.
- t_1 is opened with no error, but builder expects 2 more tables to finalize building
- the opening of t_2 causes and error, and builder returns the fully completed `PgTablesQueryResult` object
- the waiter object notifies the listener with result object
- the result for t_3 is returned (success or failure)
- at this point waiter might notify listener for second time this is not desired

To avoid multiple notifications of listener in situation above waiter resets `listener_` pointer after the first notification.
But this method is totally wrong because the `Waiter::TableReady` method is called from multiple threads.

To solve the issue the method the `TablesResultBuilder::CompleteResult` method is renamed into `TablesResultBuilder::CompleteResultOnce`. And not it returns non-null pointer only in case the result is fully built by applying of current `Result<TableInfo>` object. I.e. in case the builder has already built the `PgTablesQueryResult` object each next call of the `TablesResultBuilder::CompleteResultOnce` method will return null.
Jira: DB-17700

Test Plan:
New unit test is introduced

```
./yb_build.sh --gtest_filter PgMiniTest.OpenTableFailureDuringPerform
```

Reviewers: sergei

Reviewed By: sergei

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D46316