[#27588] YSQL: Print reason for storage flush under debug logs
Summary:
The contents of the write buffer (`PgOperationBuffer`) in pggate is flushed for a handful of reasons:
 - Reasons external to the buffer: such as end of statement/subtransaction/transaction, change of snapshot etc.
 - Reasons internal to the buffer: multiple writes to the same row, an enqueued read operations that needs to read the contents of the buffer etc.

In large queries, it is currently quite hard to determine what caused the flush. It is often important to know in order to improve the query's performance.
This revision logs the reason for flushing the write buffer when `yb_debug_log_docdb_requests` is enabled.

This revision has the following goals:
 - Improve the explainability of storage flush requests: achieved by logging the reason for flushing
 - Improve code searchability: introducing an enum of flush reasons allows grepping for values in the enum across the code base.
 - No degradation in performance: when `yb_debug_log_docdb_requests` is turned off, the debug strings are not constructed, and no memory is allocated. This reduces any performance impact.
Jira: DB-17164

Test Plan: Run Jenkins

Reviewers: pjain, jason

Reviewed By: jason

Subscribers: svc_phabricator, jason, smishra, yql

Tags: #jenkins-trigger

Differential Revision: https://phorge.dev.yugabyte.com/D44988