[#28285] docdb: Remove check for tablet hide time when loading hidden tablets
Summary:
**Context**
We used to hide (not delete) all unused tablets after a PITR, because we wanted to allow forward restores with PITR (restoring from t=20 to t=10, then back to t > 10). Forward restores were disallowed by D19137, but we still hide all tablets, including children that we will never be able to restore to.

Then, we had an issue where if a tablet split and we did PITR to before the split, and then it split again with the same partition bounds, we would have two sets of tablets with the same partition bounds and split depth but different ids. This caused a check failure, which we tried to fix in D40314 by choosing to load the set of tablets that have the same hybrid time as the table. But this was never a condition we respected: tablets are deleted before the table, not necessarily at the same time. This means that if a table is dropped and a master leader restarts, it will not load any tablets for that table in the hidden state because their hide times are strictly before the table's hide time.

Ultimately, this prevented cloning to a time before a `DROP TABLE` after a master leader failover. The clone would fail with an error like:
```
GetTableLocationsRpc(table_id: 0000400600003000800000000000413a, max_tablets: 2147483647, require_tablets_running: 1, num_attempts: 489, timed out after deadline expired, passed 120.453s of 120.000s
```

**This diff**
This diff simplifies this behavior by deleting split children that were created after the restore time, since we will never be able to get them back anyways. This allows us to remove the code added by D40314, which makes the clone test succeed.

Also cleaned up some code in `restore_sys_catalog.cc` about tablet splitting because we create register split children atomically since D31428.
Jira: DB-17969

Test Plan:
Added a drop table to the clone master leader failover test, and added a check that the tablets are deleted after failover in another test.
```
./yb_build.sh release --cxx-test integration-tests_minicluster-snapshot-test --gtest_filter PgCloneMultiMaster.CloneAfterMasterChange
./yb_build.sh release --cxx-test tools_yb-admin-snapshot-schedule-test --gtest_filter=YbAdminSnapshotScheduleTest.RestoreAfterSplit
```

Reviewers: mhaddad, zdrudi

Reviewed By: mhaddad, zdrudi

Subscribers: ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D46361