[PLAT-18401] Avoid removing PITR config which is deleted already
Summary:
The delete PITR config errors out while trying to delete a PITR config that doesn't exist.
So, we check for the presence of the PITR task before deleting it.

Test Plan: Test restart XCluster for DB scoped replication where this erorr is seen and check whether it is successful

Reviewers: hzare

Reviewed By: hzare

Differential Revision: https://phorge.dev.yugabyte.com/D46346