[#28382] DocDB: Async writes - Part 1
Summary:
Async Writes helps reduce the latency of YSQL write statements (and reads that grab locks) when they are run as part of a distributed transaction.

**Distributed transactions and the need for async writes: **
In Yugabyte, we use distributed transactions when we need to perform atomic updates to multiples tablets, or when YSQL transaction has multiple statements involved that cannot be grouped together into one write batch.
A simple example of this would be:
```
BEGIN;
INSERT INTO tbl1 VALUES (1, 'a');
SELECT COUNT(*) FROM tbl1;
INSERT INTO tbl2 VALUES(2, 'b');
INSERT INTO tbl3 VALUES(3, 'c');
COMMIT;
```

This transaction would result in 4 quorum round trips (1 for each write and 1 for the COMMIT) since on each write we wait for the RAFT quorum commit on each tablet. On a multi region stretch cluster this can range from 30-100ms due to the network latency between them.
With Async writes we defer the wait for the quorum commit to the final transaction COMMIT statement, which reduces the transaction latency to a 2 quorum round trips. Reads and write operations to happen in parallel to the RAFT quorum commit.

**Implementation details:**
When the tablet leader gets a write request (of a distributed transaction), it grabs locks on the corresponding keys, and enqueues a WriteOperation in the RAFT queue. Once the corresponding operation gets quorum committed in RAFT, the operation is marked as replicated. This results in the update to the IntentsDB, following which the locks are released and the client is notified of the completion of the request.
With Async writes, once the WriteOperation is enqueued into the RAFT queue, the client is notified with the OpId of the RAFT operation. The client then asynchronously tracks the completion of the write using the provided OpId. The write is replicated, and the intent is applied, the inly after the quorum commit. Unblocking reads is tracked in #28320.

The detailed sequence of steps are as follows:
1. The client calls `Batcher::CreateRpc` which uses the gFlags to set `use_async_write` on `WriteRequestPB` for YSQL distributed transactional writes.
2. In `WriteQuery::DoStartSynchronization` if the proto field is set, `WriteOperation::SetAsyncWrite` is called with a callback to return the write ack to the client with the OpId in the response.
3. When the WriteOperation is added to RAFT and an OpId is generated `WriteOperation::AddedAsPending` is invoked which invokes the callback. Once the OpId is quorum committed, it will invoke `WriteOperation::DoReplicated`.
4. The client processes the response in `Batcher::ProcessWriteResponse`, and if the response has a `async_write_op_id`, it records it in the transacction by calling `YBTransaction::RecordAsyncWrite`, and invokes `WaitForAsyncWriteRpc` asynchronously.
5. The client is now free to perform further reads and writes. Any reads to the same table is blocked till all async writes to it complete.
6. When the tablet leader receives the `WaitForAsyncWrite`, it records the rpc in the tablet peer using `TabletPeer::RegisterAsyncWriteCompletion`. If the leader looses its leadership (`TabletPeer::BecomeReplica`), or shutsdown (`TabletPeer::CompleteShutdown`), then it will send a failure status to all pending writes.
7. When the quorum commit of the OpId completes the leader peer is notified via `TabletPeer::MajorityReplicated`, and it calls `NotifyCommitedAsyncWrites` where all pending callbacks equal or prior to majority replicated OpId are notified of successful write.
8. When the client receives the completion ack from `WaitForAsyncWrite`, it removes the OpId from the session via `Batcher::RecordAsyncWriteCompletion`.
9. If the client transaction entres the COMMIT phase before the Async Write was completed then it will be blocked in `FinishTransaction` by `YBTransaction::WaitForAsyncWrites`.

If the tablet leader looses leadership for any reason before the `WaitForAsyncWriteRpc` completes then the client transaction fails.

#28383 Tracks handling of graceful leader moves and tablet splits without causing user errors.
#28384 Tracks async write of the intent before the quorum commit, and unblocking concurrent reads and writes to the same tablet without blocking behind the commit.

**Upgrade/Rollback Safety:**
The feature is gated by gFlag `ysql_enable_async_writes`.

**Perf numbers:**
Single zone TPCC: 24% improvement in Average NewOrder, and Insert latency, 50% improvement in Delete and Update latency. No degradation in throughput, Select latency or CPU, disk and memory usage.
Multi Region Sysbench Read-Write: ~ 24-40% reduction is latency of Insert, Update and Delete Ops. 51.1% improvement in overall throughput. CPU and disk usage has gone up proportionally. No significant degradation in Select latency or memory usage.

Fixes #28382
Jira: DB-18066

Test Plan: ydb  --cxx-test async_writes-test.cc

Reviewers: slingam, sergei

Reviewed By: sergei

Subscribers: svc_phabricator, yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D45343