[#28277] YSQL: Connection Manager - Add test to verify avg_wait_time_ns is updated for control pool
Summary:
D45898 added a fix to connection manager whereby the `avg_wait_time_ns` stat for the control
connection pool is also updated based on the number of authentication requests handled by the
control backends (specifically, the number of times a client is successfully attached on a route).

This diff adds a java test to verify that the `avg_wait_time_ns` field is non-zero for the control
connection pool when a new connection request is handled.
Locally, the test passes consistently, 30/30 times (as tested).

The test diff adds a test to the existing Java object `org.yb.ysqlconnmgr.TestStatsAndMetrics`. The
test name is `testControlPoolConnections`.
Jira: DB-17960

Test Plan:
`org.yb.ysqlconnmgr.TestStatsAndMetrics#testControlPoolConnections` should pass as this
is the test added in this diff.

Jenkins: all tests

Reviewers: skumar, asrinivasan, mkumar, arpit.saxena

Reviewed By: arpit.saxena

Subscribers: svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D46131