[PLAT-18449]  Fix NPE for continuous backups when lastBackup is null
Summary:
The lastBackup field is declared as a primitive long type, but in
the db it is a nullable column. When we try to query and hit a
record where last_backup is null, we get a NPE "Cannot
invoke "java.lang.Long.longValue()" because "o" is null".

In `ContinuousBackupMapper.java` we're already storing
`lastBackup` into a `Long` and then checking for null, this
diff just updates the type in `ContinuousBackupConfig.java`.

Test Plan:
Create a continuous backup configuration successfully.
Issue a successful get request for the backup configuration prior to
any backups being performed (so lastbackup is still `null`).
{F388497}

Reviewers: muthu

Reviewed By: muthu

Differential Revision: https://phorge.dev.yugabyte.com/D46400