[#28012] YSQL: Always send custom ParameterStatus packet from Postgres to CM
Summary:
Add a byte for custom flags at the end of YB ParameterStatus packet (type `'r'`). These flags will be used to tell connection manager what sort of GUC setting was done so it can take decisions based on that. This diff doesn't change any external functionality -- previously incorrect things are still incorrect. This is an internal change where we just use our new packet type for Postgres->ConnMgr communication.

Only one bit is being set and read from the flags: `YB_PARAM_STATUS_REPORT_ENABLED`. This is set for variables that have `GUC_REPORT` enabled. Currently we are using it to induce the same difference as was being done with 'r'/'S' packets. in other words we are using it to decide whether to forward the packet to external client or not.
Jira: DB-17630

Test Plan: Jenkins: all tests

Reviewers: skumar, rbarigidad, mkumar, asrinivasan, vikram.damle, #db-approvers

Reviewed By: skumar, vikram.damle, #db-approvers

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D45460