[PLAT-18384] Pass isDrConfig to the need_bootstrap call
Summary:
This diff updates the YBA UI code to pass an extra `isDrConfig` field
when making need_bootstrap calls. This is needed because bootstrapping
is always required for DR configs with automatic ddl replication.
Passing this field to the backend will allow us to determine when we're
dealing with this case.

This diff also makes the required changes to support the new bootstrap
reason introduced by D46257

Related backend diff: D46257

Test Plan:
Create a DR config and check the need_bootstrap request.
Add a database and check the need_bootstrap request.

Create an xCluster config and check the need_bootstrap request.
Add a table and check the need_bootstrap request.

Reviewers: hzare, rmadhavan, kkannan

Reviewed By: hzare

Differential Revision: https://phorge.dev.yugabyte.com/D46299