[BACKPORT 2025.1][PLAT-18421] Read tableId from sourceTableInfo instead of the toplevel tableId when working with xCluster tables
Summary:
This diff ensures we use the tableid provided in sourceTableInfo instead of the top level tableId field.
See `XClusterTableConfig.java` for context on the object shape.

The reason for this change comes from things like sequence data tables which do not have a simple UUID for their reported tableIds.
For sequence data tables we expect to see something like `<uuid>.sequences_data_for.<tableUuid>` instead.
But if we look at the tableId under sourceTableInfo from a XClusterTableConfig object, the table id is truncated and only shows the
uuid prefix dropping `.sequences_data_for.` and the tableUuid suffix.
This presents a problem if we are reading from both places and then try to compare for equality.
This diff resolves the problem by ensuring the UI reads tableId form sourceTableInfo.

There is one instance where the top level tableId field is used.
In `ReplicationUtils.tsx` we have a `augmentTablesWithXClusterDetails` function which gets used to display table information for
tables which are already part of the xCluster config. In this function we use the tableId specified in sourceTableInfo or targetTableInfo
when possible, but fall back to the top level tableId when both are not provided. This is fine because either
the table is dropped on both source and target which means the user should drop the table from the xCluster config OR
we're not able to get the table info from RPC calls to source/target universe which should be address first before trying to
make changes to the tables in an xCluster config.

Original commit: D46455

Test Plan:
Create an xCluster config with automatic ddl replication and verify we correctly working with only the truncated tableIds
provided in sourceTableInfo

Reviewers: hzare, rmadhavan

Reviewed By: hzare

Differential Revision: https://phorge.dev.yugabyte.com/D46456