[#28634] DocDB: Add time units to Prometheus metric help descriptions
Summary:
Many of our metrics do not have the time unit in their HELP description.

For example,
```
cpu_utime{exported_instance="ml-win-DIB1N:9000",metric_type="server",metric_id="yb.tabletserver"} 21682 1726695073741
```

While the underlying metric definitions do have a unit, it is not outputted to the Prometheus metrics only the JSON ones.

In this diff I add the time units to each relevant description.

Because some of the descriptions already have the unit I decided not to add the metric unit globally to all descriptions as that would just be redundant in most cases.

I left the hybrid time alones because they are not actually microseconds; we should ideally fix that in the future diffs.

Also did some tidying while I was at it.

Fixes #28634
Jira: DB-18328

Test Plan: started up a local cluster and did curl to see the metrics then searched for ones that did not have time units.

Reviewers: yyan

Reviewed By: yyan

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D46867