[#3945] YSQL: Add support for altering table constraint
Summary:
Enabled syntax:
`ALTER TABLE table_name ALTER CONSTRAINT constraint_name`
`    [ DEFERRABLE | NOT DEFERRABLE ] [ INITIALLY DEFERRED | INITIALLY IMMEDIATE ];`

In fact the related code is not changed.
The diff enables the syntax in the grammar and in the YB handler.
The main changes are in the tests.
Jira: DB-2036

Test Plan:
./yb_build.sh --java-test org.yb.pgsql.TestPgRegressForeignKey#testPgRegress
./yb_build.sh --java-test org.yb.pgsql.TestPgRegressPgForeignData#schedule
./yb_build.sh --java-test org.yb.pgsql.TestPgRegressPgTable#schedule

./yb_build.sh --cxx-test pgwrapper_pg_fkey-test --gtest_filter PgFKeyTest.DeferredConstraintReadAtTxnEnd
New test:
./yb_build.sh --cxx-test pgwrapper_pg_fkey-test --gtest_filter PgFKeyTest.AlteredDeferredConstraintReadAtTxnEnd

Reviewers: sanketh, kfranz

Reviewed By: kfranz

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D46358