[PLAT-18485] Ignore index table ids with no table detals
Summary:
**Context**
When hitting the list universe tables endpoint with
`xClusterSupportedOnly=True`, the backend filters out colocated child tables from
the table details array that gets returned to the API caller.
The issue comes when this colocated child table is an index table. The main table
still specifies the index table id in its `indexTableIDs` field so the client
thinks this table exists in the response. When the client tries to find the
index table from the from the API response, it is unable to and assigns undefined
to the index table. This causes issues because we expect the table ids to
correspond to actual tables being returned.

**Change**
This diff adds logic to ignore index table ids if we're not able to find a
matching index table from the list universe tables endpoint with
`xClusterSupportedOnly=True`. This ensures we don't look at any colocated index
tables.

Test Plan:
Create a pair of universes on db versions 2025.1.1+

On one of the universes, set up the following table

create extension vector;
CREATE TABLE glove100d (id BIGSERIAL, word varchar(100), vec vector(2));
INSERT INTO glove100d (word, vec) VALUES
('hello', '[0.1, 0.2]'),
('world', '[0.3, 0.4]');
create index on glove100d using ybhnsw (vec vector_l2_ops);

Set up xCluster DR with automatic ddl replication between the two universes.
Verify there is no issues and the colocated table is not included in the need
bootstrap call nor is it shown in the number of tables selected.
Verify there are no issues with edit table selection after config is created and
verify we're able to restart replication for tables in the config.

Reviewers: hzare, rmadhavan

Reviewed By: hzare

Subscribers: arastogi

Differential Revision: https://phorge.dev.yugabyte.com/D46543