[#27037] YSQL: Allow concurrent DDLs when object locking is enabled
Summary:
In YSQL, we restrict successful execution of concurrent DDLs on the same database
because we don't have a serialization mechanism in place to detect
and allow DDLs which can be safely executed concurrently.

To restrict concurrent DDLs, we rely on a 2-prong mechanism:

(1) Use snapshot isolation to detect conflicting writes from concurrent DDL transactions.
(2) Almost all DDLs increment the per-db catalog version at the end of the DDL transaction.

This ensures that 2 concurrent DDLs (even unrelated ones) update the catalog
version and face a write-write conflict on the catalog version. Note that two concurrent
DDLs conflict because they have overlapping `[read time, commit time]` windows and in
Snapshot Isolation we can't have two writes to the same piece of data in this
overlapping window.

There are some exceptions to this e.g., we allow concurrent `CREATE TABLE`s by
not incrementing the catalog version.

This restriction to allow one DDL at a time is a pain point in various ways:
(1) Hurts performance since there is no concurrency (unrelated DDLs can't be executed together)
(2) Introduces friction in usability (client has to ensure DDLs are not issued concurrently)

Vanilla postgres acquires object locks to serialize DDLs with other DDLs/ DMLs. Concurrent DDLs that
operate on the same object in a conflicting manner wait on each other. And once a DDL is unblocked
and acquires a lock, it:

(1) processes invalidation messages to remove cache entries that have been modified and hence are stale
(see AcceptInvalidationMessages() in lmgr.c that was enabled in commit `9891ccbf7c`). And,

(2) invalidates the CatalogSnapshot. This ensures that new catalog reads fetch the latest catalog
information (this commit adds code to reset the catalog read time for YSQL whenever Pg invalidates
the CatalogSnapshot, see `InvalidateCatalogSnapshot()`).

Note that invalidation messages are propagated along with the object lock release rpc and not subject to
the tserver-master heartbeat delay.

Without processing invalidation messages and invalidating the CatalogSnapshot, concurrent DDLs
that touch the same object and wait on each other (e.g., such as the ones in
`yb.orig.avoid_stale_catalog_reads-*` tests) could corrupt the system catalog. A DDL that waits and
acquires an object lock is supposed to read the modifications made by the DDL(s) it was
waiting on.

YSQL recently added support for object locking under the gflag
`enable_object_locking_for_table_locks`. With object locking in place,
we can serialize concurrent DDLs like Pg does.

## Implementation

This change integrates PG's CatalogSnapshot invalidation and refresh code paths with pggate to allow
concurrent DDLs to be handled like PG when object locking is enabled. All changes are guarded under the
`yb_fallback_to_legacy_catalog_read_time` GUC. The following changes are made:

(1) Whenever a read/ write operation to a catalog table is performed:
(a) save the current read time serial number.
(b) switch to the CatalogSnapshot's read time serial number.
(c) once the operation is done, switch back to the saved read time serial number.

(2) Switching to the CatalogSnapshot's read time serial number and back is currently done in pg_doc_op.cc
to keep the implementation simple. This requires that we don't buffer the operations since we don't have a
mechanism in place to send operations for different read time serial numbers in one batch request to
PgClientService.

(3) Acquire exclusive object locks on the catalog version and invalidation messages tables to ensure DDLs
are serialized one after the other when reading and modifying them. We want to ensure that we
read the latest catalog version, this is taken care of by the `GetCatalogSnapshotReadPoint` call in pg_doc_op.cc.
The catalog snapshot will be invalidated and a new one will be fetched since the catalog version table isn't
cached in system catalog (see the check for `RelationHasSysCache` in `GetNonHistoricCatalogSnapshot`).
See `YbCallNewSQLIncrementCatalogVersionHelper()`. The `yb.orig.concurrent_ddls_read_latest_cat_version` test
ensures that concurrent DDLs read the latest catalog version.

Refer #29285 for future follow-up work items in this area.

Test Plan:
Jenkins: urgent

./yb_build.sh release --java-test org.yb.pgsql.TestPgRegressIsolationObjectLocking#testPgRegress

Reviewers: smishra, patnaik.balivada, stiwary, sanketh

Reviewed By: patnaik.balivada

Subscribers: ybase, yugaware, nikhil, yql

Differential Revision: https://phorge.dev.yugabyte.com/D47729