[#28559] YSQL: move documentdb build logic
Summary:
The logic for whether to build documentdb is encoded in
YB_ENABLE_YSQL_DOCUMENTDB_EXT, which is determined at the start of
build.  It is then used in third-party-extensions/Makefile to disable
building documentdb.  This approach is not ideal as running make in
third-party-extensions/documentdb will bypass the check.

Move the logic into third-party-extensions/documentdb/Makefile so that,
if in the future, we run make in third-party-extensions/documentdb, it
works as intended.  An alternative approach is to further move the logic
into third-party-extensions/documentdb/pg_documentdb_core and similar
subdirectories, but such a use case is unlikely from the user's
perspective, and that approach may miss logical merge conflicts from
changes to third-party-extensions/documentdb/Makefile.

While at it, sort lines under DIRS.

Jira: DB-18254

Test Plan:
jenkins

Close: #28559

Reviewers: hsunder

Reviewed By: hsunder

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D46533