[#28070] YSQL: Fix hint generation for partitioned/child tables
Summary:
Plan hints are a mechanism to constrain planner choices so a specific plan
is used for a query. Yugabyte uses an enhanced version of the pg_hint_plan
extension that overcomes limitations found in the original version. The Yugabyte
hints functionality is described in [[ https://docs.yugabyte.com/preview/explore/query-1-performance/pg-hint-plan/ | Optimize YSQL queries using pg_hint_plan ]].

One limitation that was important to overcome is the inability of the
original implementation to effectively apply hints to queries with
tables that have the same name/alias. To apply hints to such queries, the Yugabyte
hint facility makes sure each table in a query has a unique alias.

The Yugabyte implementation can also generate hints from a plan. These generated
hints guarantee that they will produce the same plan for the query as
the plan they were generated from.

There is also a GUC, pg_hint_plan.yb_enable_internal_hint_test, that, if enabled,
tests if the generated hints lead to the same plan as the one found by the planner
(and from which the generated hints came).

The bug this diff addresses is the failure of YB hint generation to generate the
correct hints for a query involving partitioned (or child) tables where different partitions
have different plans. The reasons for the failure are:

1) the MergeAppend node type was not handled in the the YB hint generation code. This led
     to only 1 table appearing in the generated MergeJoin hint.
2) After fixing (1), hint generation/validation still failed because partitions
    were not assigned unique aliases as were other tables in the query.

When a partitioned table is used in a query, it is replaced by an Append or MergeAppend
node with the individual partitions as inputs. If the partitioned table is named 'prt1'
then all partitions comprising the table get the same alias 'prt1', which prevents
individual partitions from being used in hints. (The unique aliases for partitions
in a vanilla PG Explain are generated at Explain time and are not available for use
in the planner.)

The Yugabyte hint code makes sure each table in the query has a globally unique alias
that is available for use during planning (including hinting). This fix extends
this logic to partitions, and to child tables in an inheritance hierarchy.
The MergeAppend node is also processed correctly in the hint generation code.

The unique aliasing feature was introduced by commit: f2f0d6f49f84c0916bc2834ad7d577da9e3c212e.

During testing of this fix, an issue was discovered with hint generation for foreign tables : [[ https://github.com/yugabyte/yugabyte-db/issues/28620 | GH28620 ]]

With these changes, the sort key for the output of a MergeAppend/Append node will be taken from the left-most/first child, which will now have an alias that is distinct from the alias assigned to the MergeAppend/Append node itself. Before this fix, the node and its left-most child had the same alias.

Test Plan:
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressExtension'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressThirdPartyExtensionsPgHintPlan'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressPartitions'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressParallel'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInherit'

Reviewers: mtakahara, mihnea, jason, #db-approvers

Reviewed By: mtakahara

Subscribers: svc_phabricator, jason, yql

Differential Revision: https://phorge.dev.yugabyte.com/D46631